+++ server/lib/shared.js
@@ -63,6 +63,9 @@ const EMAIL_DOMAINS = [
   "yahoo.com.br", "hotmail.com.br", "outlook.com.br", "uol.com.br", "bol.com.br", "terra.com.br", "ig.com.br", "itelefonica.com.br", "r7.com", "zipmail.com.br", "globo.com", "globomail.com", "oi.com.br"
 ];
 
+/**
+ * See README.md Notes section for details behind events filtering
+ */
 const EXCLUDED_EVENTS = [
   "Attributes changed",
   "Entered segment",