+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.14
 
 - [improvement] Upgrade code to node v8.9.x LTS release
+- [improvement] Use customizable tabs for code editor
 - [bugfix] Fix failing code syntax check
 - [bugfix] Fix failing initial user load
 - [maintenance] Re-factoring of code base for better readibility and maintainability